+++ modelopt/torch/quantization/linear_attention/__init__.py
@@ -16,6 +16,8 @@
 """Numerical policies and differentiable references for linear attention."""
 
 from .config import *
+from .decode import *
+from .decode_prefill import *
 from .kda import *
 from .matmul import *
 from .prefill import *